Title: 합과 평균을 구하는 함수, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240106/합과 평균을 구하는 함수/function-that-obtains-sum-and-mean.cpp
@@ -3,8 +3,8 @@
 using namespace std;
 
 float n1, n2, n3;
-int sum;
-int avg;
+float sum;
+float avg;
 
 int main() {
     cin >> n1 >> n2 >> n3;
@@ -12,7 +12,7 @@ int main() {
     sum = floor(n1 + 0.5) + floor(n2 + 0.5) + floor(n3 + 0.5);
     avg = floor((sum / 3) + 0.5);
 
-    cout << sum << '\n' << avg;
+    cout << (int)sum << '\n' << (int)avg;
 
     return 0;
 }
